Move FastNoiseLite
