+++ src/main/java/fr/communaywen/core/dreamdim/DreamChunkGenerator.java
@@ -1,5 +1,6 @@
 package fr.communaywen.core.dreamdim;
 
+import fr.communaywen.core.utils.FastNoiseLite;
 import org.bukkit.Material;
 import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.generator.WorldInfo;
