+++ src/main/java/fr/communaywen/core/utils/FastNoiseLite.java
@@ -1,4 +1,4 @@
-package fr.communaywen.core.dreamdim;
+package fr.communaywen.core.utils;
 // MIT License
 //
 // Copyright(c) 2023 Jordan Peck (jordan.me2@gmail.com)
